requirements.txt
fastapi==0.111.0
uvicorn==0.23.1
requests==2.32.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0

openai==1.31.0
httpx==0.27.2

